Read pagemaps with multiple threads
Some binaries we are interested in have massive address spaces and
bookmark take a couple of minutes.

This commit adds multithreading to the reading of the pagemaps with the
rayon crate. I tried to optimise this code without threads but I didn't
manage to make it work:
- Reading a chunk of data with pread would not work, as we are only
interested in steps of PAGE_SIZE, which would use more memory that will
never be read
- mmap-ing /pagemap doesn't work. Even though procfs has support for
mmap-ed files
(https://elixir.bootlin.com/linux/v5.2.9/source/include/linux/fs.h#L1804)
this particular file doesn't implement it
https://elixir.bootlin.com/linux/v5.2.9/source/fs/proc/task_mmu.c#L1601.